# optimize generate and new table performance