hello, we came across this lib and trying to use it, but found some performance issue, the memory alloc count and alloc size can be largely reduse(result shows below) and speed up new maglev table and rebuild, it can also reduce go GC pressure.
please check this PR, looking forward to be accepted~~

benchmark before:
```bash
➜  go-maglev git:(optimize) ✗ go test -bench ^BenchmarkGenerate$ -benchmem -count 10
goos: linux
goarch: amd64
pkg: github.com/dgryski/go-maglev
BenchmarkGenerate-4            1        1118862920 ns/op        656387072 B/op       126 allocs/op
BenchmarkGenerate-4            6         211422894 ns/op        656387120 B/op       126 allocs/op
BenchmarkGenerate-4            7         146772075 ns/op        656387072 B/op       126 allocs/op
BenchmarkGenerate-4            7         146420866 ns/op        656387085 B/op       126 allocs/op
BenchmarkGenerate-4            7         146834448 ns/op        656387072 B/op       126 allocs/op
BenchmarkGenerate-4            7         147924688 ns/op        656387072 B/op       126 allocs/op
BenchmarkGenerate-4            7         146545748 ns/op        656387072 B/op       126 allocs/op
BenchmarkGenerate-4            7         146761895 ns/op        656387072 B/op       126 allocs/op
BenchmarkGenerate-4            7         146550326 ns/op        656387085 B/op       126 allocs/op
BenchmarkGenerate-4            7         146920101 ns/op        656387085 B/op       126 allocs/op
PASS
ok      github.com/dgryski/go-maglev    12.493s
```


after:
```bash

➜  go-maglev git:(optimize) ✗ go test -bench ^BenchmarkGenerate$ -benchmem -count 10
goos: linux
goarch: amd64
pkg: github.com/dgryski/go-maglev
BenchmarkGenerate-4       237949              4957 ns/op            2048 B/op          2 allocs/op
BenchmarkGenerate-4       235502              4960 ns/op            2048 B/op          2 allocs/op
BenchmarkGenerate-4       237808              5040 ns/op            2048 B/op          2 allocs/op
BenchmarkGenerate-4       236378              5012 ns/op            2048 B/op          2 allocs/op
BenchmarkGenerate-4       237189              5022 ns/op            2048 B/op          2 allocs/op
BenchmarkGenerate-4       234795              5065 ns/op            2048 B/op          2 allocs/op
BenchmarkGenerate-4       232723              4965 ns/op            2048 B/op          2 allocs/op
BenchmarkGenerate-4       234796              4964 ns/op            2048 B/op          2 allocs/op
BenchmarkGenerate-4       235342              4963 ns/op            2048 B/op          2 allocs/op
BenchmarkGenerate-4       236899              4974 ns/op            2048 B/op          2 allocs/op
PASS
ok      github.com/dgryski/go-maglev    12.560s

➜  go-maglev git:(optim_new) ✗ go test -bench ^BenchmarkNew$ -benchmem -count 10
goos: linux                                      
goarch: amd64                                                                     
pkg: github.com/dgryski/go-maglev
BenchmarkNew-4               271           4398043 ns/op          535703 B/op          5 allocs/op
BenchmarkNew-4               272           4377443 ns/op          535696 B/op          5 allocs/op
BenchmarkNew-4               273           4385998 ns/op          535696 B/op          5 allocs/op
BenchmarkNew-4               272           4381675 ns/op          535696 B/op          5 allocs/op
BenchmarkNew-4               273           4440517 ns/op          535697 B/op          5 allocs/op
BenchmarkNew-4               259           4587298 ns/op          535697 B/op          5 allocs/op
BenchmarkNew-4               259           4558834 ns/op          535697 B/op          5 allocs/op     
BenchmarkNew-4               268           4528645 ns/op          535697 B/op          5 allocs/op
BenchmarkNew-4               271           4524732 ns/op          535696 B/op          5 allocs/op
BenchmarkNew-4               254           4487935 ns/op          535697 B/op          5 allocs/op
PASS                                                    
ok      github.com/dgryski/go-maglev    16.726s  
```