@dgryski 2020-06-08
This looks nice.  I'm a bit concerned that the lookup values might be changing (especially given the one test we had for a specific permutation has been altered.). Can you add a test to make sure that the values stay consistent before and after?  I'd hate to have a lookup change when somebody upgrades this package.

Otherwise, this is good to merge.