@trainyao 2020-06-10
yes, it's better to have a unit-test to test that permutation array stays the same