@trainyao 2020-06-11
@dgryski  unit-test is added, and I also run in master branch, the unit-test can pass also in master branch.